Wind Charge Fix
+ Fixed a bug where barrels were destroyed upon contact with a wind charge explosion / wind burst.
+ Updated Spigot API to 1.21 for needed EntityTypes
+ Fixed various stuff that changed with 1.20.5 (mostly particles)

NOTE: This works (at least for 1.21+, I haven't tested older versions), but it seems to be incompatible with the recipes addon now. I'm a little unsure whether or not older server software might be affected by these changes, so I'll be moving this to another branch to play around with further if necessary before finally merging it into master.